day 05

+++ philo/philo.h
@@ -6,7 +6,7 @@
 /*   By: ebakchic <ebakchic@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/14 19:51:40 by ebakchic          #+#    #+#             */
-/*   Updated: 2023/01/22 20:55:01 by ebakchic         ###   ########.fr       */
+/*   Updated: 2023/01/22 23:35:15 by ebakchic         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct t_infop
 	int				m_eat;
 	long int		start;
 	long int		t_die;
-	int				s_sleep;
+	long int		s_sleep;
 	int				t_sleep;
 	int				t_eat;
 	long int		l_meal;
